Create README.md
Forest fire may be defined as an unclosed and freely spreading combustion that consumes the natural fuels. When a fire burns out of control it is known as Wild Fire.
The main adverse impact of the uncontrollable forest fire includes damage to growing stock of forests, loss of biodiversity, increase in soil erosion
 scorching of soil and reduction in its permeability and water retaining capacity and volatization of the nutrients like Nitrogen. 
Extreme forest fire burns thousand of houses and kills many human beings and cattle throughout the world.